Refactor tests in Python
